+++ public/index.html
@@ -9,6 +9,7 @@
   <link rel="preconnect" href="https://fonts.googleapis.com" />
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
   <link href="https://fonts.googleapis.com/css2?family=Lobster&display=swap" rel="stylesheet" />
+  <link rel="shortcut icon" href="https://raw.githubusercontent.com/ChoiRamsey/zinwoos/main/favicon.ico" />
   <title>ZINWOOS</title>
 </head>
