+++ src/components/Nav/Nav.scss
@@ -23,9 +23,7 @@
     z-index: 9999;
 
     .nav-logo {
-      width: 100px;
-      height: 100%;
-      background-color: white;
+      width: 120px;
 
       img {
         width: 100%;
